l10/1 - auth checks + logout

+++ items.php
@@ -2,8 +2,12 @@
 include 'header.php'; ?>
 
 <h1>Grocery List</h1>
-<a href="item-details.php">Add an Item</a>
 <?php
+    session_start(); // only call once per page
+    if (!empty($_SESSION['username'])) {
+        echo '<a href="item-details.php">Add an Item</a>';
+    }
+
 try {
     // 1. Connect to the db.  Host: 172.31.22.43, DB: dbNameHere, Username: usernameHere, PW: passwordHere
     include 'db.php';
@@ -21,19 +25,37 @@ try {
 
     // 5. Use a foreach loop to iterate (cycle) through all the values in the $items variable.  Inside this loop, use an echo command to display the name of each item.  See https://www.php.net/manual/en/control-structures.foreach.php for details.
     // start an HTML table for formatting BEFORE the foreach loop
-    echo '<table class="table table-striped table-light"><thead><th>Category<th>Name</th><th>Quantity</th></TH><th>Actions</th></thead>';
+    echo '<table class="table table-striped table-light">
+        <thead><th>Category</th><th>Name</th><th>Quantity</th>';
+
+    if (!empty($_SESSION['username'])) {
+        echo '<th>Actions</th>';
+    }
+
+    echo '</thead>';
 
     foreach ($items as $indItems) {
         // must use "return" to evaluate the confirm method to decide if the link should fire or not
-        echo '<tr><td>' . $indItems['category'] . '</td>
-            <td><a href="item-details.php?itemId=' . $indItems['itemId'] .
-            '">' . $indItems['name'] . '</a></td>
-            <td>' . $indItems['quantity'] . '</td>
-            <td><a href="item-details.php?itemId=' . $indItems['itemId'] .
-            '" class="btn btn-secondary">Edit</a>&nbsp;
+        echo '<tr><td>' . $indItems['category'] . '</td>';
+
+        if (!empty($_SESSION['username'])) {
+            echo '<td><a href="item-details.php?itemId=' . $indItems['itemId'] .
+                '">' . $indItems['name'] . '</a></td>';
+        }
+        else {
+            echo '<td>' . $indItems['name'] . '</td>';
+        }
+
+        echo '<td>' . $indItems['quantity'] . '</td>';
+
+        if (!empty($_SESSION['username'])) {
+            echo '<td><a href="item-details.php?itemId=' . $indItems['itemId'] .
+                '" class="btn btn-secondary">Edit</a>&nbsp;
                 <a href="delete-item.php?itemId=' . $indItems['itemId'] .
-            '" class="btn btn-danger" title="Delete"
-                onclick="return confirmDelete();">Delete</a></td></tr>';
+                '" class="btn btn-danger" title="Delete"
+                onclick="return confirmDelete();">Delete</a></td>';
+        }
+        echo '</tr>';
     }
 
     // close the table
